manager/requirements.txt
Flask~=3.0.2
flask-cors~=4.0.0
Flask-SocketIO~=5.3.6
python-socketio~=5.11.1
websocket-client~=1.7.0
gunicorn~=21.2.0
requests~=2.31.0